# stms-backend/openapi.yaml
openapi: 3.0.3
info:
  title: STMS - Smart Trainer Management System API
  version: 1.0.0
  description: |
    Authentication & User Management API for Sprint 1.
    Multi-tenant athletics management platform with club-scoped data isolation.
  contact:
    name: STMS Team
    email: dev@stms.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.stms.example.com/v1
    description: Production
  - url: https://api-staging.stms.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 signed JWT access token (15 min expiry).
        Include in Authorization header: `Bearer <access_token>`
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HttpOnly, Secure, SameSite=Strict refresh token cookie (7 days)
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors (400 only)
...
    PaginatedResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: array
          items:
            type: object
        meta:
          type: object
          properties:
            page:
              type: integer
              minimum: 1
            limit:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
            totalPages:
              type: integer

    # ==================== AUTH ====================
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        firebaseUid:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [system_admin, club_admin, coach, athlete]
        clubIds:
          type: array
          items:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        activeClubId:
          type: string
          nullable: true
          pattern: '^[0-9a-fA-F]{24}$'
        status:
          type: string
          enum: [active, invited, deactivated, deleted]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [club_admin, coach, athlete]
          description: Cannot assign system_admin via this endpoint
    InviteUserRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [club_admin, coach, athlete]
        clubId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: Defaults to inviter's activeClubId
    InviteResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            invitationId:
              type: string
            magicLink:
              type: string
            expiresAt:
              type: string
              format: date-time

    # ==================== CLUB ====================
...
  parameters:
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in current club
      description: Club admin only. Paginated, filterable.
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [club_admin, coach, athlete]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, invited, deactivated]
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires club_admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Users]
      summary: Invite user to club
      description: Club admin only. Sends magic link invitation email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '201':
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires club_admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already member of club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Club admin can view any user in club. Users can view own profile.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to view this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Users]
      summary: Update user role (club admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires club_admin; cannot change own role; cannot assign system_admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Users]
      summary: Soft delete user (club admin only)
      description: Marks user as deleted, anonymizes PII, revokes tokens
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '204':
          description: User deactivated
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires club_admin; cannot delete self
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== CLUBS ====================
...